Place ExperimentRunner in bark_ml and introduce max. vel.
