+++ bark_ml/experiment/experiment_runner.py
@@ -3,28 +3,11 @@ import os
 import hashlib
 import logging
 from pathlib import Path
-from absl import app
-from absl import flags
 import tensorflow as tf
 import numpy as np
 from bark.runtime.commons.parameters import ParameterServer
-from experiments.experiment import Experiment
+from bark_ml.experiment.experiment import Experiment
 
-FLAGS = flags.FLAGS
-flags.DEFINE_enum("mode",
-                  "visualize",
-                  ["train", "visualize", "evaluate", "print", "save"],
-                  "Mode the configuration should be executed in.")
-# NOTE: absolute paths are required
-flags.DEFINE_string("exp_json",
-                    "/Users/hart/Development/bark-ml/experiments/configs/highway_gat_small.json",
-                    "Path to the experiment json.")
-flags.DEFINE_string("save_path",
-                    "/Users/hart/Development/bark-ml/experiments/configs/new_exp.json",
-                    "Path to the experiment json.")
-flags.DEFINE_integer("random_seed",
-                     0,
-                     "Random seed to be used.")
 
 class ExperimentRunner:
   """The ExperimentRunner-Class provides an easy-to-use interface to
@@ -134,13 +117,3 @@ class ExperimentRunner:
 
   def SaveExperiment(self, file_path):
     self._params.Save(file_path)
-
-# run experiment
-def run_experiment(argv):
-  exp_runner = ExperimentRunner(
-    json_file=FLAGS.exp_json,
-    mode=FLAGS.mode,
-    random_seed=FLAGS.random_seed)
-
-if __name__ == '__main__':
-  app.run(run_experiment)
