+++ bark_ml/experiment/py_experiment_tests.py
@@ -9,14 +9,14 @@
 
 import unittest
 import os
-from experiments.experiment import Experiment
+from bark_ml.experiment.experiment import Experiment
 
 
 class PyExperimentTests(unittest.TestCase):
   def test_experiment_class(self):
-    # experiment = Experiment(
-    #   os.path.join(os.path.dirname(__file__),
-    #   "configs/highway_gnn.json"))
+    experiment = Experiment(
+      os.path.join(os.path.dirname(__file__),
+      "configs/highway_gnn.json"))
     # visualize/evaluate
     # experiment.runner.Run(num_episodes=1, render=False)
     # experiment.params["ML"]["BehaviorTFAAgents"]["CheckpointPath"] = \
